refactor: date 형식 변경 메소드 util로 분리
